Collision_avoidance_car
This a simple code for making the car move on its own by avaoiding
obstacles such as car when the car is in motion. This was implemented
using Arduino Uno and ultra sonic sensors were used to measure the
distances from all the four sides. The car moves or stops with respect
to the obstacles.